Rename id-tag column to id_tag; pin relation id dtype; add regression tests
Follow-up polish to #234:

- Rename the column that surfaces a colliding OSM `id` tag from `tagged_id` to `id_tag` (reads clearly as 'the value of the id tag').

- Expose explode_way_tags as cpdef so the id-tag handling is unit-testable (no behaviour/perf change; pbfreader still cimports it at C level).

- Restore explicit dtype=np.int64 on parsed relation ids for Windows/int64 consistency.

- Add regression tests: relation ids are read as-is and stay int64 (#170), and a way tag keyed 'id' surfaces as id_tag without overwriting the element id (#233).

+++ pyrosm/pbfreader.pyx
@@ -319,7 +319,7 @@ cdef parse_relations(data, string_table, unix_time_filter):
     # Version
     versions = np.array([rel.info.version for rel in data], dtype=np.int64)
 
-    ids = np.array([rel.id for rel in data])
+    ids = np.array([rel.id for rel in data], dtype=np.int64)
 
     timestamps = np.array([rel.info.timestamp for rel in data], dtype=np.int64)
